added temporal sorting of midi notes
